fix(validation): make scope evaluation reproducible (#212)
* fix(validation): make scope evaluation reproducible

* fix(validation): compare resolved Windows paths
