+++ .github/workflows/scope-evaluator.yml
@@ -1,0 +1,29 @@
+name: Scope evaluator
+
+on:
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  reproduce:
+    runs-on: ubuntu-latest
+    timeout-minutes: 45
+
+    steps:
+      - name: Check out full history
+        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
+        with:
+          fetch-depth: 0
+
+      - name: Set up Python 3.14.3
+        uses: actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97 # v7.0.0
+        with:
+          python-version: "3.14.3"
+
+      - name: Prepare pinned repositories
+        run: python validation/v1.0-scope-related-context/prepare.py
+
+      - name: Recompute the canonical result
+        run: python validation/v1.0-scope-related-context/evaluate.py --check
