+++ validation/v1.0-scope-related-context/REPORT.md
@@ -11,7 +11,11 @@ or an LLM-based retrieval layer.
 
 - Baseline: `e5329b29d884d40dc7d3b14ca7fa26c83a14878b`
 - Current: `ab98e39910d9dc70eacdad3fcb4df4d5e7b579de`
-- Corpus: the six v0.4 ranking holdouts and six v0.4 edge-IDF holdouts
+- Corpus: the six v0.4 ranking holdouts and six v0.4 edge-IDF holdouts. The tracked
+  [ranking manifest](corpus/v0.4-ranking-holdout/holdout.json) has SHA-256
+  `6fe09278638ccede6e4be981fcc8b2fd5fedcd9288dbcc780210032bce616736`. The tracked
+  [edge-IDF manifest](corpus/v0.4-edge-idf-holdout/holdout.json) has SHA-256
+  `1b45b645362a26a24e3d89805282b4dfd14c583b527ebe326698fbce4f0b5eaf`.
 - Selection: each manifest's fixed primary node; ranking was not run
 - Comparison: the full canonical one-hop edge set determined the affected tasks. The product review
   API's capped proposal set was measured separately for required-neighbor Recall@10 and the 10-item
@@ -22,6 +26,9 @@ or an LLM-based retrieval layer.
   either product output.
 - Runtime: CPython 3.14.3. Baseline and current were each repeated with hash seeds 0 and 1 in
   `-S` workers.
+- Measurement code: [frozen_evaluator.py](frozen_evaluator.py), preserved byte for byte from the
+  original study with SHA-256
+  `124d0e7e68c5a3352a47611ab0ea165a0373b3b832646f956d48035093238d78`.
 
 The evaluator reads canonical source snapshots from regular Git blobs at each manifest commit, so a
 checkout's symlink and line-ending policy cannot change the measured graph. It separately requires a
@@ -29,6 +36,52 @@ clean tracked worktree and rejects any collected Python file that is untracked o
 HEAD blob. For the audit only, CRLF bytes are accepted when their sole normalization to LF matches the
 blob exactly. Existing `.silobrief` files are not inputs; their complete path-and-byte digest, Git
 status, and tracked-file digest must remain unchanged across all workers.
+
+## Reproduce from a clean clone
+
+Git and CPython 3.14.3 are required. The preparation step needs network access. The evaluator itself
+uses only the prepared checkouts and Git objects.
+
+```text
+git clone https://github.com/d3vksy/silobrief.git
+cd silobrief
+python validation/v1.0-scope-related-context/prepare.py
+python validation/v1.0-scope-related-context/evaluate.py --check
+```
+
+The preparation command fetches each commit listed below and checks it out in detached HEAD state.
+It does not follow a branch or select the latest commit. Running the command again reuses a clean,
+correct checkout. It stops if a checkout has local changes or a different origin URL.
+
+| Suite | Repository URL | Commit | Checkout path |
+| --- | --- | --- | --- |
+| ranking | `https://github.com/encode/starlette.git` | `99b7cc62e0c2cb8d24e5f546b7e34e17496c265e` | `corpus/v0.4-ranking-holdout/repos/starlette` |
+| ranking | `https://github.com/vitalik/django-ninja.git` | `134869b74b6cba214284faa9f13d54b7247362c0` | `corpus/v0.4-ranking-holdout/repos/django-ninja` |
+| ranking | `https://github.com/litestar-org/litestar.git` | `ec6f9c95443d323a8cee4d9ff5dd1bd9862f5fc8` | `corpus/v0.4-ranking-holdout/repos/litestar` |
+| ranking | `https://github.com/pallets/jinja.git` | `5ef70112a1ff19c05324ff889dd30405b1002044` | `corpus/v0.4-ranking-holdout/repos/jinja` |
+| ranking | `https://github.com/pytest-dev/pytest.git` | `28e86a6c2ae0173831e4925a4af89b02a2936d09` | `corpus/v0.4-ranking-holdout/repos/pytest` |
+| ranking | `https://github.com/python-poetry/poetry.git` | `92b74dcfe348d0e01e14d40d6c1fa47a4ee04a54` | `corpus/v0.4-ranking-holdout/repos/poetry` |
+| edge-IDF | `https://github.com/pallets/flask.git` | `06ea505ce2b2042af26e96d35ebf159af7c0869d` | `corpus/v0.4-edge-idf-holdout/repos/flask` |
+| edge-IDF | `https://github.com/aio-libs/aiohttp.git` | `0ed510d8dc9f6397878fe775ac7e2fbd0b13641f` | `corpus/v0.4-edge-idf-holdout/repos/aiohttp` |
+| edge-IDF | `https://github.com/sanic-org/sanic.git` | `a64dc641a8e4ad777a4602d2bdec53d736901472` | `corpus/v0.4-edge-idf-holdout/repos/sanic` |
+| edge-IDF | `https://github.com/Textualize/rich.git` | `9cb198944f8184df92217efc8b20d3fffa4b4fa0` | `corpus/v0.4-edge-idf-holdout/repos/rich` |
+| edge-IDF | `https://github.com/psf/black.git` | `fa72105efac5c15c3a3c83c21ec6e6097c525325` | `corpus/v0.4-edge-idf-holdout/repos/black` |
+| edge-IDF | `https://github.com/pydantic/pydantic.git` | `a54529f06204ecb8f940d2506ff66dd1521917c8` | `corpus/v0.4-edge-idf-holdout/repos/pydantic` |
+
+The checkout paths are relative to `validation/v1.0-scope-related-context`. Git ignores the
+downloaded repositories, but tracks both manifests. To put the corpus elsewhere, pass the same
+`--external-root PATH` to `prepare.py` and `evaluate.py`.
+
+The evaluator never clones, fetches, checks out, or writes an external repository. It checks the
+commit and working tree before evaluation, then compares repository state after all workers finish.
+The manual [Scope evaluator workflow](../../.github/workflows/scope-evaluator.yml) runs the same two
+commands from a full-history checkout.
+
+Successful verification prints:
+
+```text
+canonical_sha256=0c65bd3a7d64bc5218583b08c332029e384bda63bcd42337999c92ffd0171923
+```
 
 ## Results
 
@@ -69,14 +122,16 @@ generated result was interpreted.
 Run from the repository root:
 
 ```text
-python validation/v1.0-scope-related-context/evaluate.py
+python validation/v1.0-scope-related-context/prepare.py
 python validation/v1.0-scope-related-context/evaluate.py --check
+python -m unittest tests.test_scope_evaluator
 python -m unittest tests.test_index tests.test_boundary_placeholders tests.test_source_review tests.test_python_structure
 ```
 
-The first command wrote the canonical result. The second independently recomputed it and returned
-the same SHA-256. The focused lexical-scope and boundary regression gate ran 64 tests successfully.
-The complete test suite also passed 234 tests with 7 expected skips, and Ruff reported no findings.
+The first command prepares only the pinned inputs. The second independently recomputes the result
+and compares it with [results.json](results.json). The original focused lexical-scope and boundary
+regression gate ran 64 tests successfully. The complete study test suite also passed 234 tests with
+7 expected skips, and Ruff reported no findings.
 
 ## Not now
