+++ validation/v1.0-scope-related-context/corpus/.gitignore
@@ -1,0 +1,1 @@
+*/repos/
